pyproject.toml
[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "OneBot V12", module_name = "nonebot.adapters.onebot.v12" },
    { name = "RedProtocol", module_name = "nonebot.adapters.red" }
]
plugins = []
plugin_dirs = ["march7th"]
builtin_plugins = []

[tool.poetry]
name = "march7th"
version = "0.2.0"
description = "March7th - Honkai Star Rail bot based on Nonebot2."
authors = ["mobyw <mobyw66@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/Mar-7th/March7th"
repository = "https://github.com/Mar-7th/March7th"
documentation = "https://github.com/Mar-7th/March7th#readme"

[tool.poetry.dependencies]
python = ">=3.8, <4.0"
nb-cli = "^1.2.6"
nonebot2 = {extras = ["fastapi", "httpx", "websockets"], version = "^2.1.1"}
nonebot-adapter-onebot = "^2.3.1"
nonebot-adapter-red = "^0.7.4"
nonebot-plugin-datastore = "^1.1.2"
nonebot-plugin-apscheduler = "^0.3.0"
nonebot-plugin-send-anything-anywhere = "^0.4.0"
pydantic = "^1.10.0"
pillow = "^10.0.1"
pil-utils = "^0.1.8"
qrcode = "^7.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
nonemoji = "^0.1.4"
pre-commit = "^3.2.2"

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.3"
pytest-asyncio = "^0.21.0"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["plugins"]
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "E501", "C901"]
line-length = 88
target-version = "py38"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
